ServerSide/serverapi/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-ninja==0.22.2
idna==3.4
numpy==1.26.1
pandas==2.1.1
pydantic==1.10.13
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
mysqlclient==2.1.0